long

+++ c_gpio/pins.c
@@ -126,20 +126,20 @@ int get_level_time_by_id(int level_id)
 void level_sleep(int level_id)
 {
 	int sleep0, sleep1;
-	float tmp;
+	long tmp;
 
 	sleep0 = get_level_time_by_id(level_id);
 	sleep1 = get_level_time_by_id(level_id + 1);
 	
 	if (sleep1 == 0)
-		tmp = T/ (float) sleep0;
+		tmp = T/ (long) sleep0;
 	else
-		tmp = T / (float) sleep0 - T / (float) sleep1;
+		tmp = T / (long) sleep0 - T / (long) sleep1;
 	
-	level_usleep(tmp*1000);
+	level_usleep(tmp * 1000);
 }
 
-void level_usleep(int us)
+void level_usleep(long us)
 {
 	struct timespec wait;
